medretainn-main/backend/requirements.txt
# MedRetain CRM - Backend Requirements
# Python 3.11 compatible

# FastAPI and Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# Data Processing

# Machine Learning
scikit-learn==1.5.0
joblib==1.3.2

# Environment and Configuration
python-dotenv==1.0.1

# Messaging
twilio==8.13.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP Client
httpx==0.26.0

# Date/Time utilities
python-dateutil==2.8.2

# Encryption and Database for HIMS
cryptography==42.0.5
